fix: wolf image 교체
